+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Added new symbols for borehole types 'virtual borehole' and 'other'.
+- Panel to edit lithostratigraphy layers.
 
 ### Changed
 
@@ -14,6 +15,7 @@
 ### Fixed
 
 - Fixed bug where empty strings were displayed instead of _unknown_ in the _casing_ dropdown when adding water ingresses, hydrotests, groundwater measurements or field measurements.
+- Fixed label and setting for lithology top bedrock attribute of layer.
 
 ## v2.0.422 - 2023-06-02